requirements.txt
torch>=2.1.0
transformers>=4.35.0
accelerate>=0.24.0
gradio>=4.0.0
spaces>=0.19.0
pyyaml>=6.0
numpy>=1.24.0
packaging>=21.0
